Refactor decorators and improve type hints
- Updated type hints in `xcom_pull`, `xcom_push`, and `task` decorators for better clarity and type safety.
- Added type ignoring comments for wrapper metadata assignments to suppress type checker warnings.
- Enhanced the `AsyncTaskQueue` and `TaskQueue` classes to ensure consistent return types by converting responses to dictionaries.
- Improved error handling in Lua script executions by ensuring that raw results are properly checked and formatted.
- Refactored the `ScheduledTask` model to use more specific type hints for the `func` attribute.
- Updated the `TaskScheduler` class to ensure synchronous Redis client creation and added type assertions for Redis operations.
- Enhanced worker classes to improve type hints and ensure better handling of task completion.
- Added comprehensive tests for `FairQueueConfig` to validate configuration loading and serialization.
- Adjusted various test cases to ensure they align with the new configuration structure and worker handling.

+++ tests/performance/test_queue_throughput.py
@@ -66,7 +66,7 @@ class TestQueueThroughput:
         num_threads = 4
         tasks_per_thread = 250
 
-        def push_tasks(thread_id: int) -> int:
+        def push_tasks(thread_id: int) -> float:
             """Push tasks from a single thread."""
             tasks = create_test_tasks(
                 tasks_per_thread,
